backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
openai>=2.6.0
python-dotenv==1.0.0
requests==2.31.0
httpx>=0.28.1
PyMySQL==1.1.0
Flask-SQLAlchemy==3.0.5
PySocks==1.7.1
